+++ Tests/Unit/SharePointDsc/SharePointDsc.SPSearchTopology.Tests.ps1
@@ -47,35 +47,43 @@ Describe "SPSearchTopology - SharePoint Build $((Get-Item $SharePointCmdletModul
             }
         }
 
-        Add-Type -TypeDefinition "public class AdminComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-        Add-Type -TypeDefinition "public class CrawlComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-        Add-Type -TypeDefinition "public class ContentProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-        Add-Type -TypeDefinition "public class AnalyticsProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-        Add-Type -TypeDefinition "public class QueryProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-        Add-Type -TypeDefinition "public class IndexComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;} public System.Int32 IndexPartitionOrdinal {get; set;}}"
+        Add-Type -TypeDefinition "public class AdminComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;} public System.Guid ServerId {get; set;}}"
+        Add-Type -TypeDefinition "public class CrawlComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;} public System.Guid ServerId {get; set;}}"
+        Add-Type -TypeDefinition "public class ContentProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;} public System.Guid ServerId {get; set;}}"
+        Add-Type -TypeDefinition "public class AnalyticsProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;} public System.Guid ServerId {get; set;}}"
+        Add-Type -TypeDefinition "public class QueryProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;} public System.Guid ServerId {get; set;}}"
+        Add-Type -TypeDefinition "public class IndexComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;} public System.Int32 IndexPartitionOrdinal {get; set;} public System.Guid ServerId {get; set;}}"
+
+        $serverId = New-Guid
 
         $adminComponent = New-Object AdminComponent
         $adminComponent.ServerName = $env:COMPUTERNAME
+        $adminComponent.ServerId = $serverId
         $adminComponent.ComponentId = [Guid]::NewGuid()
 
         $crawlComponent = New-Object CrawlComponent
         $crawlComponent.ServerName = $env:COMPUTERNAME
+        $crawlComponent.ServerId = $serverId
         $crawlComponent.ComponentId = [Guid]::NewGuid()
 
         $contentProcessingComponent = New-Object ContentProcessingComponent
         $contentProcessingComponent.ServerName = $env:COMPUTERNAME
+        $contentProcessingComponent.ServerId = $serverId
         $contentProcessingComponent.ComponentId = [Guid]::NewGuid()
 
         $analyticsProcessingComponent = New-Object AnalyticsProcessingComponent
         $analyticsProcessingComponent.ServerName = $env:COMPUTERNAME
+        $analyticsProcessingComponent.ServerId = $serverId
         $analyticsProcessingComponent.ComponentId = [Guid]::NewGuid()
 
         $queryProcessingComponent = New-Object QueryProcessingComponent
         $queryProcessingComponent.ServerName = $env:COMPUTERNAME
+        $queryProcessingComponent.ServerId = $serverId
         $queryProcessingComponent.ComponentId = [Guid]::NewGuid()
 
         $indexComponent = New-Object IndexComponent
         $indexComponent.ServerName = $env:COMPUTERNAME
+        $indexComponent.ServerId = $serverId
         $indexComponent.IndexPartitionOrdinal = 0
 
         Mock Start-SPEnterpriseSearchServiceInstance { return $null }
@@ -88,6 +96,15 @@ Describe "SPSearchTopology - SharePoint Build $((Get-Item $SharePointCmdletModul
         Mock New-SPEnterpriseSearchIndexComponent { return @{} }
         Mock Set-SPEnterpriseSearchTopology { return @{} }
         Mock Remove-SPEnterpriseSearchComponent { return $null }
+
+        Mock Get-SPServer {
+            return @(
+                @{
+                    Name = $env:COMPUTERNAME
+                    Id = $serverId
+                }
+            )
+        }
 
         Context "No search topology has been applied" {
             Mock Get-SPEnterpriseSearchComponent {
@@ -206,30 +223,29 @@ Describe "SPSearchTopology - SharePoint Build $((Get-Item $SharePointCmdletModul
                 return @($adminComponent, $crawlComponent, $contentProcessingComponent, $analyticsProcessingComponent, $queryProcessingComponent)
             }
             Mock Get-SPEnterpriseSearchComponent {
-                Add-Type -TypeDefinition "public class AdminComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-                Add-Type -TypeDefinition "public class CrawlComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-                Add-Type -TypeDefinition "public class ContentProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-                Add-Type -TypeDefinition "public class AnalyticsProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-                Add-Type -TypeDefinition "public class QueryProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-
                 $adminComponent = New-Object AdminComponent
                 $adminComponent.ServerName = $env:COMPUTERNAME
+                $adminComponent.ServerId = $serverId
                 $adminComponent.ComponentId = [Guid]::NewGuid()
 
                 $crawlComponent = New-Object CrawlComponent
                 $crawlComponent.ServerName = $env:COMPUTERNAME
+                $crawlComponent.ServerId = $serverId
                 $crawlComponent.ComponentId = [Guid]::NewGuid()
 
                 $contentProcessingComponent = New-Object ContentProcessingComponent
                 $contentProcessingComponent.ServerName = $env:COMPUTERNAME
+                $contentProcessingComponent.ServerId = $serverId
                 $contentProcessingComponent.ComponentId = [Guid]::NewGuid()
 
                 $analyticsProcessingComponent = New-Object AnalyticsProcessingComponent
                 $analyticsProcessingComponent.ServerName = $env:COMPUTERNAME
+                $analyticsProcessingComponent.ServerId = $serverId
                 $analyticsProcessingComponent.ComponentId = [Guid]::NewGuid()
 
                 $queryProcessingComponent = New-Object QueryProcessingComponent
                 $queryProcessingComponent.ServerName = $env:COMPUTERNAME
+                $queryProcessingComponent.ServerId = $serverId
                 $queryProcessingComponent.ComponentId = [Guid]::NewGuid()
 
                 return @($adminComponent, $crawlComponent, $contentProcessingComponent, $analyticsProcessingComponent, $queryProcessingComponent)
@@ -248,35 +264,34 @@ Describe "SPSearchTopology - SharePoint Build $((Get-Item $SharePointCmdletModul
                 return @($adminComponent, $crawlComponent, $contentProcessingComponent, $analyticsProcessingComponent, $queryProcessingComponent, $indexComponent)
             }
             Mock Get-SPEnterpriseSearchComponent {
-                Add-Type -TypeDefinition "public class AdminComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-                Add-Type -TypeDefinition "public class CrawlComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-                Add-Type -TypeDefinition "public class ContentProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-                Add-Type -TypeDefinition "public class AnalyticsProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-                Add-Type -TypeDefinition "public class QueryProcessingComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;}}"
-                Add-Type -TypeDefinition "public class IndexComponent { public string ServerName { get; set; } public System.Guid ComponentId {get; set;} public System.Int32 IndexPartitionOrdinal {get; set;}}"
-
                 $adminComponent = New-Object AdminComponent
                 $adminComponent.ServerName = $env:COMPUTERNAME
+                $adminComponent.ServerId = $serverId
                 $adminComponent.ComponentId = [Guid]::NewGuid()
 
                 $crawlComponent = New-Object CrawlComponent
                 $crawlComponent.ServerName = $env:COMPUTERNAME
+                $crawlComponent.ServerId = $serverId
                 $crawlComponent.ComponentId = [Guid]::NewGuid()
 
                 $contentProcessingComponent = New-Object ContentProcessingComponent
                 $contentProcessingComponent.ServerName = $env:COMPUTERNAME
+                $contentProcessingComponent.ServerId = $serverId
                 $contentProcessingComponent.ComponentId = [Guid]::NewGuid()
 
                 $analyticsProcessingComponent = New-Object AnalyticsProcessingComponent
                 $analyticsProcessingComponent.ServerName = $env:COMPUTERNAME
+                $analyticsProcessingComponent.ServerId = $serverId
                 $analyticsProcessingComponent.ComponentId = [Guid]::NewGuid()
 
                 $queryProcessingComponent = New-Object QueryProcessingComponent
                 $queryProcessingComponent.ServerName = $env:COMPUTERNAME
+                $queryProcessingComponent.ServerId = $serverId
                 $queryProcessingComponent.ComponentId = [Guid]::NewGuid()
 
                 $indexComponent = New-Object IndexComponent
                 $indexComponent.ServerName = $env:COMPUTERNAME
+                $indexComponent.ServerId = $serverId
                 $indexComponent.IndexPartitionOrdinal = 0
 
                 return @($adminComponent, $crawlComponent, $contentProcessingComponent, $analyticsProcessingComponent, $queryProcessingComponent, $indexComponent)